+++ src/solidPod.test.ts
@@ -32,10 +32,12 @@ import {
   mockFetchError,
   overwriteFile,
   createContainerAt,
+  setThing,
 } from "@inrupt/solid-client";
 import { WithResourceInfo } from "@inrupt/solid-client/src/interfaces";
 import { INRUPT_TEST } from "@inrupt/vocab-inrupt-test-rdfdatafactory";
 import { updateOrInsertResourceInSolidPod } from "./solidPod";
+import { buildDataset } from "./solidDatasetUtil";
 
 // Ideally we'd have a mock WebID per test suite, but it causes conflicts when
 // running all test suites. Would like to refactor to remove the duplication
@@ -135,6 +137,33 @@ function mockSaveSolidDatasetAt<Dataset extends SolidDataset>(
 }
 
 describe("Solid Pod functions", () => {
+  describe("Multiple Things in a single dataset", () => {
+    it("should fail if more than one Thing", async () => {
+      const thing1 = buildThing()
+        .addIri(INRUPT_TEST.somePredicate, INRUPT_TEST.somePodResource)
+        .build();
+      const thing2 = buildThing()
+        .addIri(
+          INRUPT_TEST.someOtherPredicate,
+          INRUPT_TEST.someOtherPodResource
+        )
+        .build();
+      const dataset = setThing(buildDataset(thing1), thing2);
+
+      const authnModule = jest.requireMock(
+        "@inrupt/solid-client-authn-node"
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      ) as { Session: any };
+      const session = new authnModule.Session(true);
+
+      // We can use the same dataset twice for this test, we just need 'more
+      // than one' to kick off the dataset sorting code...
+      await expect(
+        updateOrInsertResourceInSolidPod(session, [dataset, dataset])
+      ).rejects.toThrow("more than one Thing");
+    });
+  });
+
   describe("Inserting or updating blob with metadata into Pod", () => {
     it("should insert Blob with no metadata", async () => {
       const obj = { hello: "world" };
@@ -172,9 +201,11 @@ describe("Solid Pod functions", () => {
       const blob = new Blob([JSON.stringify(obj, null, 2)], {
         type: "application/json",
       });
-      const metadata = buildThing()
-        .addIri(INRUPT_TEST.somePredicate, INRUPT_TEST.somePodResource)
-        .build();
+      const metadata = buildDataset(
+        buildThing()
+          .addIri(INRUPT_TEST.somePredicate, INRUPT_TEST.somePodResource)
+          .build()
+      );
 
       jest
         .spyOn(
@@ -217,9 +248,11 @@ describe("Solid Pod functions", () => {
       const blob = new Blob([JSON.stringify(obj, null, 2)], {
         type: "application/json",
       });
-      const metadata = buildThing()
-        .addIri(INRUPT_TEST.somePredicate, INRUPT_TEST.somePodResource)
-        .build();
+      const metadata = buildDataset(
+        buildThing()
+          .addIri(INRUPT_TEST.somePredicate, INRUPT_TEST.somePodResource)
+          .build()
+      );
 
       jest
         .spyOn(
@@ -275,8 +308,8 @@ describe("Solid Pod functions", () => {
       const session = new authnModule.Session(true);
 
       const result = await updateOrInsertResourceInSolidPod(session, [
-        thing1,
-        thing2,
+        buildDataset(thing1),
+        buildDataset(thing2),
       ]);
       expect(result).toContain(
         "Successfully inserted or updated [2] resources"
@@ -294,7 +327,9 @@ describe("Solid Pod functions", () => {
       ) as { Session: any };
       const session = new authnModule.Session(false);
 
-      const result = await updateOrInsertResourceInSolidPod(session, [thing]);
+      const result = await updateOrInsertResourceInSolidPod(session, [
+        buildDataset(thing),
+      ]);
       expect(result).toContain("session is not logged in");
 
       // Attempt insert with number of resources that's not 1 (so 'plural').
@@ -375,7 +410,9 @@ describe("Solid Pod functions", () => {
           .addIri(INRUPT_TEST.somePredicate, INRUPT_TEST.somePodResource)
           .build();
 
-        const result = await updateOrInsertResourceInSolidPod(session, [thing]);
+        const result = await updateOrInsertResourceInSolidPod(session, [
+          buildDataset(thing),
+        ]);
         expect(result).toContain("Successfully inserted or updated [1]");
         expect(result).toContain(`[${mockWebId}]`);
       }