+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The following changes have been implemented but not released yet:
   collection.
 - Added Hobby vocabs and ETL.
 - Added util functions for describing a collection of resources.
+- Updated to treat resources as SolidDatasets instead of just Things.
 
 ## [0.0.1] - 2022-02-01